course_work4/requirements.txt
Flask~=2.1.2
flask-restx~=0.5.1
Flask-Cors~=3.0.10
Flask-SQLAlchemy~=2.5.1
marshmallow~=3.16.0
PyJWT~=2.4.0

pytest~=7.1.2
SQLAlchemy~=1.4.39
Werkzeug~=2.1.2